backend/requirements.txt
asgiref==3.10.0
attrs==25.4.0
autobahn==25.10.2
Automat==25.4.16
certifi==2025.10.5
cffi==2.0.0
channels==4.3.1
channels_redis==4.3.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
constantly==23.10.4
cryptography==46.0.3
daphne==4.2.1
Django==5.2.7
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
h11==0.16.0
hyperlink==21.0.0
idna==3.11
incremental==24.7.2
msgpack==1.1.2
psycopg==3.2.12
psycopg-binary==3.2.12
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
PyJWT==2.10.1
pyOpenSSL==25.3.0
python-decouple==3.8
python-dotenv==1.2.1
redis==7.0.1
requests==2.32.5
service-identity==24.2.0
setuptools==80.9.0
sqlparse==0.5.3
Twisted==25.5.0
txaio==25.9.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
websocket-client==1.9.0
whitenoise==6.11.0
zope.interface==8.0.1

# Email with provider abstraction
django-anymail[sendgrid]==10.2
jinja2==3.1.4

# Async task queue
celery==5.4.0